Merge pull request #120 from Gameaday/copilot/resolve-build-issues

feat: app modernization — module-boundary architecture, domain interfaces, resource decoupling, getSync() deprecation fixes, Phase D/E extension + resource decoupling, spotless lint fixes, multi-module compile fixes, Koin compile safety cleanup, and ...

+++ feature/settings/src/main/kotlin/ephyra/feature/settings/screen/SettingsBrowseScreenModel.kt
@@ -10,7 +10,7 @@ class SettingsBrowseScreenModel(
     private val getExtensionRepoCount: GetExtensionRepoCount,
 ) : ScreenModel {
 
-    fun getExtensionRepoCount(): Flow<Long> {
+    fun getExtensionRepoCount(): Flow<Int> {
         return getExtensionRepoCount.subscribe()
     }
 }
